+++ src/story-patch.js
@@ -15,8 +15,46 @@
  */
 import {minifyInlineJs} from './utils/minify-inline-js';
 
-export const navigationPatch = minifyInlineJs(`
-  if (window.history && window.history.replaceState) {
-    window.history.replaceState({ampStoryPageId: '$pageId$'}, '');
-  }
+const httpsCircumventionPatch = minifyInlineJs(`
+  (doc => {
+    const createElement = doc.createElement;
+    doc.createElement = function(tagName) {
+      const el = createElement.apply(doc, arguments);
+      if (/^a$/i.test(tagName)) {
+        Object.defineProperty(el, 'protocol', {value: 'https:'});
+      }
+      return el;
+    };
+  })(document);
+  `);
+
+const blobCorsPatch = minifyInlineJs(`
+  (win => {
+    const realFetch = win.fetch;
+    win.fetch = function(url) {
+      const args = Array.prototype.slice.call(arguments);
+      if (url.startsWith('blob:')) {
+        args[0] = url.split('?')[0];
+      }
+      return realFetch.apply(win, args);
+    }
+  })(window);
 `);
+
+const addContentToHead = (docStr, headContent) =>
+  docStr.replace('<head>', `<head><script>${headContent}</script>`);
+
+const insertHttpsCircumventionPatch = docStr =>
+  addContentToHead(docStr, httpsCircumventionPatch);
+
+const insertBlobCorsPatch = docStr => addContentToHead(docStr, blobCorsPatch);
+
+/**
+ * Patches an <amp-story>  document string for REPL support:
+ * - Monkey-patches `document.createElement()` to circumvent AMP's HTTPS checks.
+ * - Monkey-patches `window.fetch()` to allow blob url without adding amp cors param.
+ * @param {string} docStr
+ * @return {string}
+ */
+export const patchStoryDoc = docStr =>
+  insertHttpsCircumventionPatch(insertBlobCorsPatch(docStr));
